@@ -28,7 +28,7 @@
         }
     },
     "require": {
-        "php": ">=7.4.0",
+        "php": ">=8.0",
         "ext-pdo": "*",
         "ext-json": "*",
         "swayok/utils": "^1.7",
