@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace PeskyORM\Core;
 
+use PDO;
 use PDOStatement;
 
 interface DbAdapterInterface
@@ -29,38 +30,36 @@ interface DbAdapterInterface
      * @param null|string $code - callback code to prevent duplicate usage
      * @return static
      */
-    public function onConnect(\Closure $callback, ?string $code = null);
+    public function onConnect(\Closure $callback, ?string $code = null): static;
     
-    /**
-     * @return static
-     */
-    public function disconnect();
+    public function disconnect(): static;
     
     /**
      * Get last executed query
-     * @return null|string
      */
     public function getLastQuery(): ?string;
     
     /**
-     * @param string|DbExpr $query
-     * @return int|array = array: returned if $returning argument is not empty
+     * @return int - affected rows count
      */
-    public function exec($query);
+    public function exec(string|DbExpr $query): int;
     
     /**
      * @param string|DbExpr $query
      * @param array $options - see PDO::prepare()
-     * @return int|array = array: returned if $returning argument is not empty
+     * @return \PDOStatement
      */
-    public function prepare($query, array $options = []): PDOStatement;
+    public function prepare(string|DbExpr $query, array $options = []): \PDOStatement;
     
     /**
      * @param string|DbExpr $query
-     * @param string|null $fetchData - how to fetch data (one of DbAdapter::FETCH_*)
-     * @return \PDOStatement|array|string|null|int|bool|float
+     * @param string $fetchData - how to fetch data (one of DbAdapter::FETCH_*)
+     * @return mixed
      */
-    public function query($query, string $fetchData = DbAdapter::FETCH_STATEMENT);
+    public function query(
+        string|DbExpr $query,
+        string $fetchData = DbAdapter::FETCH_STATEMENT
+    ): mixed;
     
     /**
      * Listen for DB notifications (mostly for PostgreSQL LISTEN...NOTIFY)
@@ -69,22 +68,23 @@ interface DbAdapterInterface
      *      function(string $payload): boolean { return true; } - if it returns false: listener will stop
      * @param int $sleepIfNoNotificationMs - miliseconds to sleep if there were no notifications last time
      * @param int $sleepAfterNotificationMs - miliseconds to sleep after notification consumed
-     * @return void
      */
-    public function listen(string $channel, \Closure $handler, int $sleepIfNoNotificationMs = 1000, int $sleepAfterNotificationMs = 0): void;
+    public function listen(
+        string $channel,
+        \Closure $handler,
+        int $sleepIfNoNotificationMs = 1000,
+        int $sleepAfterNotificationMs = 0
+    ): void;
     
     /**
      * Set DB timezone for current session
-     * @param string $timezone
-     * @return static
      */
-    public function setTimezone(string $timezone);
+    public function setTimezone(string $timezone): static;
     
     /**
      * @param string $newSearchPath - coma-separated list of DB schemas
-     * @return static
      */
-    public function setSearchPath(string $newSearchPath);
+    public function setSearchPath(string $newSearchPath): static;
     
     /**
      * @param string $table
@@ -99,7 +99,13 @@ interface DbAdapterInterface
      * @param string $pkName - Name of primary key for $returning in DB drivers that support only getLastInsertId()
      * @return array|null - array returned only if $returning is not empty
      */
-    public function insert(string $table, array $data, array $dataTypes = [], $returning = false, string $pkName = 'id'): ?array;
+    public function insert(
+        string $table,
+        array $data,
+        array $dataTypes = [],
+        bool|array $returning = false,
+        string $pkName = 'id'
+    ): ?array;
     
     /**
      * @param string $table
@@ -115,7 +121,14 @@ interface DbAdapterInterface
      * @param string $pkName - Name of primary key for $returning in DB drivers that support only getLastInsertId()
      * @return array|null - array returned only if $returning is not empty
      */
-    public function insertMany(string $table, array $columns, array $data, array $dataTypes = [], $returning = false, string $pkName = 'id'): ?array;
+    public function insertMany(
+        string $table,
+        array $columns,
+        array $data,
+        array $dataTypes = [],
+        bool|array $returning = false,
+        string $pkName = 'id'
+    ): ?array;
     
     /**
      * @param string $table
@@ -132,7 +145,13 @@ interface DbAdapterInterface
      *          - int: number of modified rows (when $returning === false)
      *          - array: modified records (when $returning !== false)
      */
-    public function update(string $table, array $data, $conditions, array $dataTypes = [], $returning = false);
+    public function update(
+        string $table,
+        array $data,
+        string|DbExpr $conditions,
+        array $dataTypes = [],
+        bool|array $returning = false
+    ): array|int;
     
     /**
      * @param string $table
@@ -143,40 +162,30 @@ interface DbAdapterInterface
      *          - array: list of columns names to return values for
      * @return array|int - int: number of deleted records | array: returned only if $returning is not empty
      */
-    public function delete(string $table, $conditions, $returning = false);
+    public function delete(string $table, string|DbExpr $conditions, bool|array $returning = false): array|int;
     
     public function inTransaction(): bool;
     
     /**
      * @param bool $readOnly - true: transaction only reads data
      * @param null|string $transactionType - type of transaction
-     * @return static
      */
-    public function begin(bool $readOnly = false, ?string $transactionType = null);
+    public function begin(bool $readOnly = false, ?string $transactionType = null): static;
+    
+    public function commit(): static;
+    
+    public function rollBack(): static;
     
     /**
-     * @return static
+     * @param PDOStatement|PDO|null $pdoStatement $pdoStatement - if null: static->getConnection() will be used
      */
-    public function commit();
-    
-    /**
-     * @return static
-     */
-    public function rollBack();
-    
-    /**
-     * @param null|\PDOStatement|\PDO $pdoStatement $pdoStatement - if null: static->getConnection() will be used
-     * @return array
-     */
-    public function getPdoError($pdoStatement = null): array;
+    public function getPdoError(null|PDOStatement|PDO $pdoStatement = null): array;
     
     /**
      * Quote DB entity name (column, table, alias, schema)
-     * @param string $name
      * Names format:
      *  1. 'table', 'column', 'TableAlias'
      *  2. 'TableAlias.column' - quoted like '`TableAlias`.`column`'
-     * @return string
      */
     public function quoteDbEntityName(string $name): string;
     
@@ -184,24 +193,21 @@ interface DbAdapterInterface
      * Test if $name matches a DB entity naming rules or it is a JSON selector
      * @param string $name
      * @param bool $canBeAJsonSelector - test if $name contains a JSON selector like 'col_name -> json_key'
-     * @return bool
      */
     public static function isValidDbEntityName(string $name, bool $canBeAJsonSelector = true): bool;
     
     /**
      * Quote passed value
-     * @param string|int|float|bool|array|DbExpr|AbstractSelect|null $value
+     * @param string|int|float|bool|array|AbstractSelect|DbExpr|null $value
      * @param int|null $valueDataType - one of \PDO::PARAM_* or null for autodetection (detects bool, null, string only)
-     * @return string
      */
-    public function quoteValue($value, ?int $valueDataType = null): string;
+    public function quoteValue(string|int|float|bool|array|DbExpr|AbstractSelect|null $value, ?int $valueDataType = null): string;
     
     public function quoteDbExpr(DbExpr $expression): string;
     
     /**
      * Does DB supports table schemas?
      * Postgres - yes: "public"."table_name"; Mysql - no
-     * @return bool
      */
     public function isDbSupportsTableSchemas(): bool;
     
@@ -211,31 +217,29 @@ interface DbAdapterInterface
      */
     public function getDefaultTableSchema(): ?string;
     
-    /**
-     * @param string $operator
-     * @param array|float|int|string|bool|null|DbExpr $value
-     * @return string
-     */
-    public function convertConditionOperator(string $operator, $value): string;
+    public function convertConditionOperator(
+        string $operator,
+        string|int|float|bool|array|DbExpr|AbstractSelect|null $value
+    ): string;
     
     /**
      * Assemble value for condition
-     * @param string|array|int|float|bool|DbExpr|null $value
-     * @param string $operator
-     * @param bool $valueAlreadyQuoted
-     * @return string
      */
-    public function assembleConditionValue($value, string $operator, bool $valueAlreadyQuoted = false): string;
+    public function assembleConditionValue(
+        string|int|float|bool|array|DbExpr|AbstractSelect|null $value,
+        string $operator,
+        bool $valueAlreadyQuoted = false
+    ): string;
     
     /**
      * Assemble condition from prepared parts
-     * @param string $quotedColumn
-     * @param string $operator
-     * @param string|array|int|float|bool|DbExpr|null $rawValue
-     * @param bool $valueAlreadyQuoted
-     * @return string
      */
-    public function assembleCondition(string $quotedColumn, string $operator, $rawValue, bool $valueAlreadyQuoted = false): string;
+    public function assembleCondition(
+        string $quotedColumn,
+        string $operator,
+        string|int|float|bool|array|DbExpr|AbstractSelect|null $rawValue,
+        bool $valueAlreadyQuoted = false
+    ): string;
     
     /**
      * Converts general representation of data type conversion to adapter's specific one
@@ -255,7 +259,6 @@ interface DbAdapterInterface
      * @param string $table
      * @param array $columns - empty array means "all columns" (SELECT *), must contain only strings and DbExpr objects
      * @param DbExpr|null $conditionsAndOptions - Anything to add to query after "FROM $table"
-     * @return array
      */
     public function select(string $table, array $columns = [], ?DbExpr $conditionsAndOptions = null): array;
     
@@ -266,9 +269,8 @@ interface DbAdapterInterface
      * @param string $table
      * @param string|DbExpr $column
      * @param DbExpr|null $conditionsAndOptions - Anything to add to query after "FROM $table"
-     * @return array
      */
-    public function selectColumn(string $table, $column, ?DbExpr $conditionsAndOptions = null): array;
+    public function selectColumn(string $table, string|DbExpr $column, ?DbExpr $conditionsAndOptions = null): array;
     
     /**
      * Select many records form DB by compiling simple query from passed parameters returning an associative array.
@@ -277,9 +279,13 @@ interface DbAdapterInterface
      * @param string|DbExpr $keysColumn
      * @param string|DbExpr $valuesColumn
      * @param DbExpr|null $conditionsAndOptions - Anything to add to query after "FROM $table"
-     * @return array
      */
-    public function selectAssoc(string $table, $keysColumn, $valuesColumn, ?DbExpr $conditionsAndOptions = null): array;
+    public function selectAssoc(
+        string $table,
+        string|DbExpr $keysColumn,
+        string|DbExpr $valuesColumn,
+        ?DbExpr $conditionsAndOptions = null
+    ): array;
     
     /**
      * Select first matching record form DB by compiling simple query from passed parameters.
@@ -287,7 +293,6 @@ interface DbAdapterInterface
      * @param string $table
      * @param array $columns - empty array means "all columns" (SELECT *), must contain only strings and DbExpr objects
      * @param DbExpr|null $conditionsAndOptions - Anything to add to query after "FROM $table"
-     * @return array
      */
     public function selectOne(string $table, array $columns = [], ?DbExpr $conditionsAndOptions = null): array;
     
@@ -297,9 +302,8 @@ interface DbAdapterInterface
      * @param string $table
      * @param DbExpr $expression - something like "COUNT(*)" or anything else
      * @param DbExpr|null $conditionsAndOptions - Anything to add to query after "FROM $table"
-     * @return string|null|int|bool|float
      */
-    public function selectValue(string $table, DbExpr $expression, ?DbExpr $conditionsAndOptions = null);
+    public function selectValue(string $table, DbExpr $expression, ?DbExpr $conditionsAndOptions = null): mixed;
     
     /**
      * Make a simple SELECT query from passed parameters
@@ -314,7 +318,6 @@ interface DbAdapterInterface
      * Get table description from DB
      * @param string $table
      * @param null|string $schema - name of DB schema that contains $table (for PostgreSQL)
-     * @return TableDescription
      */
     public function describeTable(string $table, ?string $schema = null): TableDescription;
     
@@ -322,7 +325,6 @@ interface DbAdapterInterface
      * Search for $table in $schema
      * @param string $table
      * @param null|string $schema - name of DB schema that contains $table (for PostgreSQL)
-     * @return bool
      */
     public function hasTable(string $table, ?string $schema = null): bool;
     
@@ -330,7 +332,6 @@ interface DbAdapterInterface
      * Return DbExpr to set default value for a column.
      * Example for MySQL and PostgreSQL: DbExpr::create('DEFAULT') and used for updates and inserts
      * Note: throw exception if adapter does not support this feature
-     * @return DbExpr
      */
     public static function getExpressionToSetDefaultValueForAColumn(): DbExpr;
     
