@@ -213,14 +213,14 @@ abstract class Table implements TableInterface
     /**
      * Select associative array
      * Note: does not support columns from foreign models
-     * @param string|DbExpr $keysColumn
-     * @param string|DbExpr $valuesColumn
+     * @param string|DbExpr|null $keysColumn
+     * @param string|DbExpr|null $valuesColumn
      * @param array $conditions
      * @param \Closure|null $configurator - closure to configure OrmSelect. function (OrmSelect $select) {}
      * @return array
      * @throws \PDOException
      */
-    public static function selectAssoc($keysColumn, $valuesColumn, array $conditions = [], ?\Closure $configurator = null): array
+    public static function selectAssoc(string|DbExpr|null $keysColumn, string|DbExpr|null $valuesColumn, array $conditions = [], ?\Closure $configurator = null): array
     {
         return static::makeSelect([], $conditions, $configurator)
             ->fetchAssoc($keysColumn, $valuesColumn);
