@@ -5,9 +5,6 @@ namespace PeskyORM\ORM;
 interface ValueToObjectConverterInterface
 {
     
-    /**
-     * @return static
-     */
-    public static function createObjectFromArray(array $data);
+    public static function createObjectFromArray(array $data): static;
     
 }
