@@ -118,8 +118,8 @@ class MysqlAdapterGeneralFunctionalityTest extends BaseTestCase
     
     public function testQuotingOfInvalidIntDbValue2(): void
     {
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage("\$value expected to be integer or numeric string. Object fo class [\PeskyORM\Adapter\Mysql] received");
+        $this->expectException(\TypeError::class);
+        $this->expectExceptionMessage('Argument #1 ($value) must be of type PeskyORM\Core\DbExpr|PeskyORM\Core\AbstractSelect|array|string|int|float|bool|null');
         $adapter = static::getValidAdapter();
         /** @noinspection PhpParamsInspection */
         $adapter->quoteValue($adapter, PDO::PARAM_INT);
@@ -135,8 +135,8 @@ class MysqlAdapterGeneralFunctionalityTest extends BaseTestCase
     
     public function testQuotingOfInvalidIntDbValue4(): void
     {
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage("\$value expected to be integer or numeric string.");
+        $this->expectException(\TypeError::class);
+        $this->expectExceptionMessage('Argument #1 ($value) must be of type PeskyORM\Core\DbExpr|PeskyORM\Core\AbstractSelect|array|string|int|float|bool|null');
         $adapter = static::getValidAdapter();
         $adapter->quoteValue(curl_init('http://test.url'), PDO::PARAM_INT);
     }
